[daemon/lib/asset]Implement RehashingAsset as a kindof Readonly asset that immediately generates hashIds for all it's content.

+++ daemon/cache/map.d
@@ -245,7 +245,7 @@ public:
     /************************************************************************************
      * Add a segment into the cachemap
      ***********************************************************************************/
-    void add(ulong start, uint length) {
+    void add(ulong start, ulong length) {
         // Just skip 0-length segments
         if (!length) return;
